requirements.txt
asgiref==3.7.2
Django==5.0
django-filter==23.5
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
